Refactor project and dashboard components to use a new card-repo component for improved code organization and maintainability; adjust styles and transitions for better user experience.
