+++ src/app/app.component.scss
@@ -17,7 +17,7 @@ main{
     min-width: 250px;
     user-select: none;
     border-right: 1px solid var(--mat-sys-outline-variant);
-    transition: all 0.2s ease;
+    transition: all 0.3s ease;
 
     ul{
         list-style-type: none;
