+++ src/app/components/card-repo/card-repo.component.html
@@ -1,0 +1,38 @@
+
+<div class="flex-group gap-16 flex-wrap">
+    <div 
+        tabindex="0" 
+        class="card" 
+        *ngFor="let item of repos" 
+        (click)="open($event, item)"
+        (mousemove)="onMouseMove($event)"
+        (mouseenter)="onMouseEnter(item)"
+        (mouseleave)="onMouseLeave()"
+        >
+        <span *ngIf="item.unread > 0 " class="new">{{item.unread}}</span>
+        <mat-icon>favorite_border</mat-icon>
+        <div class="card-header">
+            <h2>{{item.name}}</h2>
+        </div>
+        <div class="card-body">
+            <p>{{item.description}}</p>
+        </div>
+    </div>
+
+    <div class="repository-container pt-48">
+        <button (click)="close()" class="close-button"><mat-icon>close</mat-icon></button>
+        <div class="flex-group flex-column gap-8 pt-8">
+            <markdown [data]="selectedRepo?.readme"></markdown>
+        </div>
+    </div>
+
+    <div #hoverContainer class="hover-container" *ngIf="isHovered" [ngStyle]="{ 
+        'top': mouseY + 'px', 
+        'left': hoverDirection === 'right' ? (mouseX + 20) + 'px' : (mouseX - (containerWidth + 16)) + 'px',
+        }">
+        <mat-spinner *ngIf="isLoading"></mat-spinner>
+        <div *ngIf="!isLoading" class="hover-card">
+            <markdown [data]="readme"></markdown>
+        </div>
+    </div>
+</div>
