+++ src/app/components/card-repo/card-repo.component.scss
@@ -1,0 +1,109 @@
+.card{
+    padding: 24px;
+    transition: all 0.1s ease;
+    border-radius: 0.5rem;
+    flex: 1 1 200px;
+    position: relative;
+    cursor: pointer;
+    max-width: 250px;
+    min-height: 180px;
+    overflow: hidden;
+
+    .card-body,.card-header{
+        pointer-events: none;
+    }
+
+    .new{
+        background: var(--mat-sys-primary);
+        color: var(--mat-sys-on-primary);
+        padding: 2px 4px;
+        position: absolute;
+        top: 8px;
+        left: 8px;
+        border-radius: 0.2rem;
+        font-size: 12px;
+    }
+
+    &:hover, &.focus{
+        transform: scale(1.03);
+        border-radius: 1rem;
+        background: var(--mat-sys-outline-variant);
+    }
+
+    &.selected{
+        z-index: 100;
+        background: var(--mat-sys-background);
+    }
+
+    mat-icon{
+        position: absolute;
+        top: 8px;
+        right: 8px;
+        color: var(--mat-sys-primary);
+        display: none;
+    }
+}
+
+.repository-container{
+    background: var(--mat-sys-background);
+    height: 100%;
+    width: 100%;
+    position: fixed;
+    top: 0;
+    transform: translateX(100%);
+    transition: all 0.4s ease;
+    box-shadow: 0 4px 6px 2px rgba($color: #000000, $alpha: 0.2);
+    z-index: 100;
+    border-radius: 24px 0 0 0;
+    padding: 56px 16px 0;
+    overflow: scroll;
+    opacity: 0;
+
+    .close-button{
+        border: 1px solid var(--mat-sys-outline-variant);
+        background: transparent;
+        cursor: pointer;
+        transition: all 0.2s ease;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        padding: 4px;
+        border-radius: 4px;
+        color: var(--mat-sys-outline);
+
+
+        &:hover{
+            background: var(--mat-sys-outline-variant);
+            transform: scale(1.2);
+            border-color: transparent;
+            color: var(--mat-sys-on-surface);
+
+            mat-icon{
+                transition: transform 0.2s ease;
+                transform: rotateZ(90deg);
+            }
+        }
+    }
+}
+
+
+.hover-container{
+    position: fixed;
+    color: var(--mat-sys-on-surface);
+    /* From https://css.glass */
+    background: color-mix(in srgb, var(--mat-sys-background) 14%, transparent); //rgba(0, 0, 0, 0.14);
+    border-radius: 8px;
+    box-shadow: 0 4px 30px color-mix(in srgb, var(--mat-sys-background) 10%, transparent);;
+    backdrop-filter: blur(7px);
+    -webkit-backdrop-filter: blur(7px);
+    border: 1px solid color-mix(in srgb, var(--mat-sys-background) 18%, transparent);
+    padding: 1rem;  /* Minimum width */
+    max-width: 500px;  /* Max width (optional, for control) */
+    width: fit-content; /* Adjust based on content */
+    z-index: 1000;
+
+    mat-spinner{
+        width: 24px !important;
+        height: 24px !important;
+    }
+}
