+++ src/app/components/card-repo/card-repo.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { CardRepoComponent } from './card-repo.component';
+
+describe('CardRepoComponent', () => {
+  let component: CardRepoComponent;
+  let fixture: ComponentFixture<CardRepoComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [CardRepoComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(CardRepoComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
