+++ src/app/pages/dashboard/dashboard.component.html
@@ -1,33 +1,7 @@
 <div class="container flex-group flex-column gap-16">
     <div class="flex-group flex-column gap-8">
         <h3>Recent projects</h3>
-        <div class="flex-group gap-8 flex-wrap">
-            <ng-container *ngFor="let item of repositories">
-                <div class="card"
-                (mousemove)="onMouseMove($event)"
-                (mouseleave)="onMouseLeave()"
-                (mouseenter)="onMouseEnter(item)">
-                    <span *ngIf="item.unread > 0" class="new">{{item.unread}}</span>
-                    <mat-icon>favorite_border</mat-icon>
-                    <div class="card-header">
-                        <h2>{{item.name}}</h2>
-                    </div>
-                    <div class="card-body">
-                        <p>{{item.description}}</p>
-                    </div>
-                </div>
-            </ng-container>
-            
-            <div #hoverContainer class="hover-container" *ngIf="isHovered" [ngStyle]="{ 
-                'top': mouseY + 'px', 
-                'left': hoverDirection === 'right' ? (mouseX + 20) + 'px' : (mouseX - (containerWidth + 16)) + 'px',
-                }">
-                <mat-spinner *ngIf="isLoading"></mat-spinner>
-                <div *ngIf="!isLoading" class="hover-card">
-                    <markdown [data]="readme"></markdown>
-                </div>
-            </div>
-        </div>
+        <app-card-repo [repos]="repositories"></app-card-repo>
         <a class="align-self-end" href="#">See more...</a>
     </div>
