+++ src/app/pages/projects/projects.component.scss
@@ -9,94 +9,6 @@
     overflow: hidden;
 }
 
-.card{
-    padding: 24px;
-    transition: all 0.1s ease;
-    border-radius: 0.5rem;
-    flex: 1 1 200px;
-    position: relative;
-    cursor: pointer;
-    max-width: 250px;
-    min-height: 180px;
-    overflow: hidden;
-
-    .card-body,.card-header{
-        pointer-events: none;
-    }
-
-    .new{
-        background: var(--mat-sys-primary);
-        color: var(--mat-sys-on-primary);
-        padding: 2px 4px;
-        position: absolute;
-        top: 8px;
-        left: 8px;
-        border-radius: 0.2rem;
-        font-size: 12px;
-    }
-
-    &:hover, &.focus{
-        transform: scale(1.03);
-        border-radius: 1rem;
-        background: var(--mat-sys-outline-variant);
-    }
-
-    &.selected{
-        z-index: 100;
-        background: var(--mat-sys-background);
-    }
-
-    mat-icon{
-        position: absolute;
-        top: 8px;
-        right: 8px;
-        color: var(--mat-sys-primary);
-        display: none;
-    }
-}
-
-.repository-container{
-    background: var(--mat-sys-background);
-    height: 100%;
-    width: 100%;
-    position: fixed;
-    top: 0;
-    transform: translateX(100%);
-    transition: all 0.4s ease;
-    box-shadow: 0 4px 6px 2px rgba($color: #000000, $alpha: 0.2);
-    z-index: 100;
-    border-radius: 24px 0 0 0;
-    padding: 56px 16px 0;
-    overflow: scroll;
-    opacity: 0;
-
-    .close-button{
-        border: 1px solid var(--mat-sys-outline-variant);
-        background: transparent;
-        cursor: pointer;
-        transition: all 0.2s ease;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        padding: 4px;
-        border-radius: 4px;
-        color: var(--mat-sys-outline);
-
-
-        &:hover{
-            background: var(--mat-sys-outline-variant);
-            transform: scale(1.2);
-            border-color: transparent;
-            color: var(--mat-sys-on-surface);
-
-            mat-icon{
-                transition: transform 0.2s ease;
-                transform: rotateZ(90deg);
-            }
-        }
-    }
-}
-
 .button-toggle-container{
     button{
         background: transparent;
